+++ lib/ts/recipe/emailpassword/index.ts
@@ -22,74 +22,106 @@ export default class Wrapper {
 
     static Error = SuperTokensError;
 
-    static signUp(email: string, password: string) {
-        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.signUp({ email, password });
+    static signUp(email: string, password: string, userContext?: any) {
+        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.signUp({
+            email,
+            password,
+            userContext: userContext === undefined ? {} : userContext,
+        });
     }
 
-    static signIn(email: string, password: string) {
-        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.signIn({ email, password });
+    static signIn(email: string, password: string, userContext?: any) {
+        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.signIn({
+            email,
+            password,
+            userContext: userContext === undefined ? {} : userContext,
+        });
     }
 
-    static getUserById(userId: string) {
-        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.getUserById({ userId });
+    static getUserById(userId: string, userContext?: any) {
+        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.getUserById({
+            userId,
+            userContext: userContext === undefined ? {} : userContext,
+        });
     }
 
-    static getUserByEmail(email: string) {
-        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.getUserByEmail({ email });
+    static getUserByEmail(email: string, userContext?: any) {
+        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.getUserByEmail({
+            email,
+            userContext: userContext === undefined ? {} : userContext,
+        });
     }
 
-    static createResetPasswordToken(userId: string) {
-        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.createResetPasswordToken({ userId });
+    static createResetPasswordToken(userId: string, userContext?: any) {
+        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.createResetPasswordToken({
+            userId,
+            userContext: userContext === undefined ? {} : userContext,
+        });
     }
 
-    static resetPasswordUsingToken(token: string, newPassword: string) {
-        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.resetPasswordUsingToken({ token, newPassword });
+    static resetPasswordUsingToken(token: string, newPassword: string, userContext?: any) {
+        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.resetPasswordUsingToken({
+            token,
+            newPassword,
+            userContext: userContext === undefined ? {} : userContext,
+        });
     }
 
-    static updateEmailOrPassword(input: { userId: string; email?: string; password?: string }) {
-        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.updateEmailOrPassword(input);
+    static updateEmailOrPassword(input: { userId: string; email?: string; password?: string; userContext?: any }) {
+        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.updateEmailOrPassword({
+            userContext: {},
+            ...input,
+        });
     }
 
-    static async createEmailVerificationToken(userId: string) {
+    static async createEmailVerificationToken(userId: string, userContext?: any) {
         let recipeInstance = Recipe.getInstanceOrThrowError();
         return recipeInstance.emailVerificationRecipe.recipeInterfaceImpl.createEmailVerificationToken({
             userId,
-            email: await recipeInstance.getEmailForUserId(userId),
+            email: await recipeInstance.getEmailForUserId(userId, userContext),
+            userContext: userContext === undefined ? {} : userContext,
         });
     }
 
-    static async verifyEmailUsingToken(token: string) {
+    static async verifyEmailUsingToken(token: string, userContext?: any) {
         let recipeInstance = Recipe.getInstanceOrThrowError();
         let response = await recipeInstance.emailVerificationRecipe.recipeInterfaceImpl.verifyEmailUsingToken({
             token,
+            userContext: userContext === undefined ? {} : userContext,
         });
         if (response.status === "OK") {
-            return await recipeInstance.recipeInterfaceImpl.getUserById({ userId: response.user.id });
+            return await recipeInstance.recipeInterfaceImpl.getUserById({
+                userId: response.user.id,
+                userContext: userContext === undefined ? {} : userContext,
+            });
         }
         return response;
     }
 
-    static async isEmailVerified(userId: string): Promise<boolean> {
+    static async isEmailVerified(userId: string, userContext?: any): Promise<boolean> {
         let recipeInstance = Recipe.getInstanceOrThrowError();
         return recipeInstance.emailVerificationRecipe.recipeInterfaceImpl.isEmailVerified({
             userId,
-            email: await recipeInstance.getEmailForUserId(userId),
+            email: await recipeInstance.getEmailForUserId(userId, userContext),
+            userContext: userContext === undefined ? {} : userContext,
         });
     }
 
-    static async revokeEmailVerificationTokens(userId: string) {
+    static async revokeEmailVerificationTokens(userId: string, userContext?: any) {
         let recipeInstance = Recipe.getInstanceOrThrowError();
         return await recipeInstance.emailVerificationRecipe.recipeInterfaceImpl.revokeEmailVerificationTokens({
             userId,
-            email: await recipeInstance.getEmailForUserId(userId),
+            email: await recipeInstance.getEmailForUserId(userId, userContext),
+            userContext: userContext === undefined ? {} : userContext,
         });
     }
 
-    static async unverifyEmail(userId: string) {
+    static async unverifyEmail(userId: string, userContext?: any) {
         let recipeInstance = Recipe.getInstanceOrThrowError();
         return await recipeInstance.emailVerificationRecipe.recipeInterfaceImpl.unverifyEmail({
             userId,
-            email: await recipeInstance.getEmailForUserId(userId),
+            email: await recipeInstance.getEmailForUserId(userId, userContext),
+            userContext: userContext === undefined ? {} : userContext,
         });
     }
 }