+++ lib/ts/recipe/emailverification/types.ts
@@ -17,9 +17,9 @@ import { BaseRequest, BaseResponse } from "../../framework";
 import OverrideableBuilder from "supertokens-js-override";
 
 export type TypeInput = {
-    getEmailForUserId: (userId: string) => Promise<string>;
-    getEmailVerificationURL?: (user: User) => Promise<string>;
-    createAndSendCustomEmail?: (user: User, emailVerificationURLWithToken: string) => Promise<void>;
+    getEmailForUserId: (userId: string, userContext: any) => Promise<string>;
+    getEmailVerificationURL?: (user: User, userContext: any) => Promise<string>;
+    createAndSendCustomEmail?: (user: User, emailVerificationURLWithToken: string, userContext: any) => Promise<void>;
     override?: {
         functions?: (
             originalImplementation: RecipeInterface,
@@ -30,9 +30,9 @@ export type TypeInput = {
 };
 
 export type TypeNormalisedInput = {
-    getEmailForUserId: (userId: string) => Promise<string>;
-    getEmailVerificationURL: (user: User) => Promise<string>;
-    createAndSendCustomEmail: (user: User, emailVerificationURLWithToken: string) => Promise<void>;
+    getEmailForUserId: (userId: string, userContext: any) => Promise<string>;
+    getEmailVerificationURL: (user: User, userContext: any) => Promise<string>;
+    createAndSendCustomEmail: (user: User, emailVerificationURLWithToken: string, userContext: any) => Promise<void>;
     override: {
         functions: (
             originalImplementation: RecipeInterface,
@@ -51,6 +51,7 @@ export type RecipeInterface = {
     createEmailVerificationToken(input: {
         userId: string;
         email: string;
+        userContext: any;
     }): Promise<
         | {
               status: "OK";
@@ -61,13 +62,18 @@ export type RecipeInterface = {
 
     verifyEmailUsingToken(input: {
         token: string;
+        userContext: any;
     }): Promise<{ status: "OK"; user: User } | { status: "EMAIL_VERIFICATION_INVALID_TOKEN_ERROR" }>;
 
-    isEmailVerified(input: { userId: string; email: string }): Promise<boolean>;
+    isEmailVerified(input: { userId: string; email: string; userContext: any }): Promise<boolean>;
 
-    revokeEmailVerificationTokens(input: { userId: string; email: string }): Promise<{ status: "OK" }>;
+    revokeEmailVerificationTokens(input: {
+        userId: string;
+        email: string;
+        userContext: any;
+    }): Promise<{ status: "OK" }>;
 
-    unverifyEmail(input: { userId: string; email: string }): Promise<{ status: "OK" }>;
+    unverifyEmail(input: { userId: string; email: string; userContext: any }): Promise<{ status: "OK" }>;
 };
 
 export type APIOptions = {
@@ -85,12 +91,14 @@ export type APIInterface = {
         | ((input: {
               token: string;
               options: APIOptions;
+              userContext: any;
           }) => Promise<{ status: "OK"; user: User } | { status: "EMAIL_VERIFICATION_INVALID_TOKEN_ERROR" }>);
 
     isEmailVerifiedGET:
         | undefined
         | ((input: {
               options: APIOptions;
+              userContext: any;
           }) => Promise<{
               status: "OK";
               isVerified: boolean;
@@ -98,5 +106,8 @@ export type APIInterface = {
 
     generateEmailVerifyTokenPOST:
         | undefined
-        | ((input: { options: APIOptions }) => Promise<{ status: "EMAIL_ALREADY_VERIFIED_ERROR" | "OK" }>);
+        | ((input: {
+              options: APIOptions;
+              userContext: any;
+          }) => Promise<{ status: "EMAIL_ALREADY_VERIFIED_ERROR" | "OK" }>);
 };