+++ lib/ts/recipe/session/types.ts
@@ -226,59 +226,83 @@ export type RecipeInterface = {
         userId: string;
         accessTokenPayload?: any;
         sessionData?: any;
+        userContext: any;
     }): Promise<SessionContainerInterface>;
 
     getSession(input: {
         req: any;
         res: any;
         options?: VerifySessionOptions;
+        userContext: any;
     }): Promise<SessionContainerInterface | undefined>;
 
-    refreshSession(input: { req: any; res: any }): Promise<SessionContainerInterface>;
+    refreshSession(input: { req: any; res: any; userContext: any }): Promise<SessionContainerInterface>;
     /**
      * Used to retrieve all session information for a given session handle. Can be used in place of:
      * - getSessionData
      * - getAccessTokenPayload
      */
-    getSessionInformation(input: { sessionHandle: string }): Promise<SessionInformation>;
+    getSessionInformation(input: { sessionHandle: string; userContext: any }): Promise<SessionInformation>;
 
-    revokeAllSessionsForUser(input: { userId: string }): Promise<string[]>;
+    revokeAllSessionsForUser(input: { userId: string; userContext: any }): Promise<string[]>;
 
-    getAllSessionHandlesForUser(input: { userId: string }): Promise<string[]>;
+    getAllSessionHandlesForUser(input: { userId: string; userContext: any }): Promise<string[]>;
 
-    revokeSession(input: { sessionHandle: string }): Promise<boolean>;
+    revokeSession(input: { sessionHandle: string; userContext: any }): Promise<boolean>;
 
-    revokeMultipleSessions(input: { sessionHandles: string[] }): Promise<string[]>;
+    revokeMultipleSessions(input: { sessionHandles: string[]; userContext: any }): Promise<string[]>;
 
-    updateSessionData(input: { sessionHandle: string; newSessionData: any }): Promise<void>;
+    updateSessionData(input: { sessionHandle: string; newSessionData: any; userContext: any }): Promise<void>;
 
-    updateAccessTokenPayload(input: { sessionHandle: string; newAccessTokenPayload: any }): Promise<void>;
+    updateAccessTokenPayload(input: {
+        sessionHandle: string;
+        newAccessTokenPayload: any;
+        userContext: any;
+    }): Promise<void>;
 
-    getAccessTokenLifeTimeMS(): Promise<number>;
+    regenerateAccessToken(input: {
+        accessToken: string;
+        newAccessTokenPayload?: any;
+        userContext: any;
+    }): Promise<{
+        status: "OK";
+        session: {
+            handle: string;
+            userId: string;
+            userDataInJWT: any;
+        };
+        accessToken?: {
+            token: string;
+            expiry: number;
+            createdTime: number;
+        };
+    }>;
 
-    getRefreshTokenLifeTimeMS(): Promise<number>;
+    getAccessTokenLifeTimeMS(input: { userContext: any }): Promise<number>;
+
+    getRefreshTokenLifeTimeMS(input: { userContext: any }): Promise<number>;
 };
 
 export interface SessionContainerInterface {
-    revokeSession(): Promise<void>;
+    revokeSession(userContext?: any): Promise<void>;
 
-    getSessionData(): Promise<any>;
+    getSessionData(userContext?: any): Promise<any>;
 
-    updateSessionData(newSessionData: any): Promise<any>;
+    updateSessionData(newSessionData: any, userContext?: any): Promise<any>;
 
-    getUserId(): string;
+    getUserId(userContext?: any): string;
 
-    getAccessTokenPayload(): any;
+    getAccessTokenPayload(userContext?: any): any;
 
-    getHandle(): string;
+    getHandle(userContext?: any): string;
 
-    getAccessToken(): string;
+    getAccessToken(userContext?: any): string;
 
-    updateAccessTokenPayload(newAccessTokenPayload: any): Promise<void>;
+    updateAccessTokenPayload(newAccessTokenPayload: any, userContext?: any): Promise<void>;
 
-    getTimeCreated(): Promise<number>;
+    getTimeCreated(userContext?: any): Promise<number>;
 
-    getExpiry(): Promise<number>;
+    getExpiry(userContext?: any): Promise<number>;
 }
 
 export type APIOptions = {
@@ -291,12 +315,13 @@ export type APIOptions = {
 };
 
 export type APIInterface = {
-    refreshPOST: undefined | ((input: { options: APIOptions }) => Promise<void>);
+    refreshPOST: undefined | ((input: { options: APIOptions; userContext: any }) => Promise<void>);
 
     signOutPOST:
         | undefined
         | ((input: {
               options: APIOptions;
+              userContext: any;
           }) => Promise<{
               status: "OK";
           }>);
@@ -304,6 +329,7 @@ export type APIInterface = {
     verifySession(input: {
         verifySessionOptions: VerifySessionOptions | undefined;
         options: APIOptions;
+        userContext: any;
     }): Promise<SessionContainerInterface | undefined>;
 };