+++ lib/ts/recipe/session/with-jwt/sessionClass.ts
@@ -28,42 +28,42 @@ export default class SessionClassWithJWT implements SessionContainerInterface {
         this.openIdRecipeImplementation = openIdRecipeImplementation;
         this.originalSessionClass = originalSessionClass;
     }
-    revokeSession = (): Promise<void> => {
-        return this.originalSessionClass.revokeSession();
+    revokeSession = (userContext?: any): Promise<void> => {
+        return this.originalSessionClass.revokeSession(userContext);
     };
-    getSessionData = (): Promise<any> => {
-        return this.originalSessionClass.getSessionData();
+    getSessionData = (userContext?: any): Promise<any> => {
+        return this.originalSessionClass.getSessionData(userContext);
     };
-    updateSessionData = (newSessionData: any): Promise<any> => {
-        return this.originalSessionClass.updateSessionData(newSessionData);
+    updateSessionData = (newSessionData: any, userContext?: any): Promise<any> => {
+        return this.originalSessionClass.updateSessionData(newSessionData, userContext);
     };
-    getUserId = (): string => {
-        return this.originalSessionClass.getUserId();
+    getUserId = (userContext?: any): string => {
+        return this.originalSessionClass.getUserId(userContext);
     };
-    getAccessTokenPayload = () => {
-        return this.originalSessionClass.getAccessTokenPayload();
+    getAccessTokenPayload = (userContext?: any) => {
+        return this.originalSessionClass.getAccessTokenPayload(userContext);
     };
-    getHandle = (): string => {
-        return this.originalSessionClass.getHandle();
+    getHandle = (userContext?: any): string => {
+        return this.originalSessionClass.getHandle(userContext);
     };
-    getAccessToken = (): string => {
-        return this.originalSessionClass.getAccessToken();
+    getAccessToken = (userContext?: any): string => {
+        return this.originalSessionClass.getAccessToken(userContext);
     };
-    getTimeCreated = (): Promise<number> => {
-        return this.originalSessionClass.getTimeCreated();
+    getTimeCreated = (userContext?: any): Promise<number> => {
+        return this.originalSessionClass.getTimeCreated(userContext);
     };
-    getExpiry = (): Promise<number> => {
-        return this.originalSessionClass.getExpiry();
+    getExpiry = (userContext?: any): Promise<number> => {
+        return this.originalSessionClass.getExpiry(userContext);
     };
 
-    updateAccessTokenPayload = async (newAccessTokenPayload: any): Promise<void> => {
+    updateAccessTokenPayload = async (newAccessTokenPayload: any, userContext?: any): Promise<void> => {
         newAccessTokenPayload =
             newAccessTokenPayload === null || newAccessTokenPayload === undefined ? {} : newAccessTokenPayload;
-        let accessTokenPayload = this.getAccessTokenPayload();
+        let accessTokenPayload = this.getAccessTokenPayload(userContext);
         let jwtPropertyName = accessTokenPayload[ACCESS_TOKEN_PAYLOAD_JWT_PROPERTY_NAME_KEY];
 
         if (jwtPropertyName === undefined) {
-            return this.originalSessionClass.updateAccessTokenPayload(newAccessTokenPayload);
+            return this.originalSessionClass.updateAccessTokenPayload(newAccessTokenPayload, userContext);
         }
 
         let existingJWT = accessTokenPayload[jwtPropertyName];
@@ -93,8 +93,9 @@ export default class SessionClassWithJWT implements SessionContainerInterface {
             userId: this.getUserId(),
             jwtPropertyName,
             openIdRecipeImplementation: this.openIdRecipeImplementation,
+            userContext,
         });
 
-        return await this.originalSessionClass.updateAccessTokenPayload(newAccessTokenPayload);
+        return await this.originalSessionClass.updateAccessTokenPayload(newAccessTokenPayload, userContext);
     };
 }