+++ lib/ts/recipe/thirdparty/types.ts
@@ -21,6 +21,7 @@ import { TypeInput as TypeInputEmailVerification } from "../emailverification/ty
 import { BaseRequest, BaseResponse } from "../../framework";
 import { NormalisedAppinfo } from "../../types";
 import OverrideableBuilder from "supertokens-js-override";
+import { SessionContainerInterface } from "../session/types";
 
 const TypeAny = {
     type: "any",
@@ -37,14 +38,18 @@ export type TypeProviderGetResponse = {
         url: string;
         params: { [key: string]: string | ((request: any) => string) };
     };
-    getProfileInfo: (authCodeResponse: any) => Promise<UserInfo>;
-    getClientId: () => string;
-    getRedirectURI?: () => string; // if undefined, the redirect_uri is set on the frontend.
+    getProfileInfo: (authCodeResponse: any, userContext: any) => Promise<UserInfo>;
+    getClientId: (userContext: any) => string;
+    getRedirectURI?: (userContext: any) => string; // if undefined, the redirect_uri is set on the frontend.
 };
 
 export type TypeProvider = {
     id: string;
-    get: (redirectURI: string | undefined, authCodeFromRequest: string | undefined) => TypeProviderGetResponse;
+    get: (
+        redirectURI: string | undefined,
+        authCodeFromRequest: string | undefined,
+        userContext: any
+    ) => TypeProviderGetResponse;
     isDefault?: boolean; // if not present, we treat it as false
 };
 
@@ -60,8 +65,8 @@ export type User = {
 };
 
 export type TypeInputEmailVerificationFeature = {
-    getEmailVerificationURL?: (user: User) => Promise<string>;
-    createAndSendCustomEmail?: (user: User, emailVerificationURLWithToken: string) => Promise<void>;
+    getEmailVerificationURL?: (user: User, userContext: any) => Promise<string>;
+    createAndSendCustomEmail?: (user: User, emailVerificationURLWithToken: string, userContext: any) => Promise<void>;
 };
 
 const InputEmailVerificationFeatureSchema = {
@@ -148,11 +153,15 @@ export type TypeNormalisedInput = {
 };
 
 export type RecipeInterface = {
-    getUserById(input: { userId: string }): Promise<User | undefined>;
+    getUserById(input: { userId: string; userContext: any }): Promise<User | undefined>;
 
-    getUsersByEmail(input: { email: string }): Promise<User[]>;
+    getUsersByEmail(input: { email: string; userContext: any }): Promise<User[]>;
 
-    getUserByThirdPartyInfo(input: { thirdPartyId: string; thirdPartyUserId: string }): Promise<User | undefined>;
+    getUserByThirdPartyInfo(input: {
+        thirdPartyId: string;
+        thirdPartyUserId: string;
+        userContext: any;
+    }): Promise<User | undefined>;
 
     signInUp(input: {
         thirdPartyId: string;
@@ -161,6 +170,7 @@ export type RecipeInterface = {
             id: string;
             isVerified: boolean;
         };
+        userContext: any;
     }): Promise<
         | { status: "OK"; createdNewUser: boolean; user: User }
         | {
@@ -188,6 +198,7 @@ export type APIInterface = {
         | ((input: {
               provider: TypeProvider;
               options: APIOptions;
+              userContext: any;
           }) => Promise<{
               status: "OK";
               url: string;
@@ -202,11 +213,13 @@ export type APIInterface = {
               authCodeResponse?: any;
               clientId?: string;
               options: APIOptions;
+              userContext: any;
           }) => Promise<
               | {
                     status: "OK";
                     createdNewUser: boolean;
                     user: User;
+                    session: SessionContainerInterface;
                     authCodeResponse: any;
                 }
               | { status: "NO_EMAIL_GIVEN_BY_PROVIDER" }
@@ -218,5 +231,5 @@ export type APIInterface = {
 
     appleRedirectHandlerPOST:
         | undefined
-        | ((input: { code: string; state: string; options: APIOptions }) => Promise<void>);
+        | ((input: { code: string; state: string; options: APIOptions; userContext: any }) => Promise<void>);
 };